Change example name; highlight Mapbox APIs usage
