--- Examples/ObjectiveC/Headers/RainFallExample.h
@@ -1,5 +1,0 @@
-#import <UIKit/UIKit.h>
-
-@interface RainFallExample : UIViewController
-
-@end
